# webpac/src/WebpacClient/Generated/Models/AuthRequest.cs
// Code generated by Microsoft (R) AutoRest Code Generator 0.16.0.0
// Changes may cause incorrect behavior and will be lost if the code is
// regenerated.

namespace Webpac.Client.Models
{
    using System;
    using System.Linq;
    using System.Collections.Generic;
    using Newtonsoft.Json;
    using Microsoft.Rest;
    using Microsoft.Rest.Serialization;

    public partial class AuthRequest
    {
        /// <summary>
        /// Initializes a new instance of the AuthRequest class.
        /// </summary>
        public AuthRequest() { }

        /// <summary>
        /// Initializes a new instance of the AuthRequest class.
        /// </summary>
        public AuthRequest(string username = default(string), string password = default(string))
        {
            Username = username;
            Password = password;
        }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "username")]
        public string Username { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "password")]
        public string Password { get; set; }

    }
}