# webpac/src/WebpacClient/Generated/Models/Authentication.cs
// Code generated by Microsoft (R) AutoRest Code Generator 0.16.0.0
// Changes may cause incorrect behavior and will be lost if the code is
// regenerated.

namespace Webpac.Client.Models
{
    using System;
    using System.Linq;
    using System.Collections.Generic;
    using Newtonsoft.Json;
    using Microsoft.Rest;
    using Microsoft.Rest.Serialization;

    public partial class Authentication
    {
        /// <summary>
        /// Initializes a new instance of the Authentication class.
        /// </summary>
        public Authentication() { }

        /// <summary>
        /// Initializes a new instance of the Authentication class.
        /// </summary>
        public Authentication(bool? authenticated = default(bool?), string user = default(string), string role = default(string), string token = default(string), DateTime? tokenExpires = default(DateTime?))
        {
            Authenticated = authenticated;
            User = user;
            Role = role;
            Token = token;
            TokenExpires = tokenExpires;
        }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "Authenticated")]
        public bool? Authenticated { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "User")]
        public string User { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "Role")]
        public string Role { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "Token")]
        public string Token { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "TokenExpires")]
        public DateTime? TokenExpires { get; set; }

    }
}