# webpac/src/WebpacClient/swagger.json
{"swagger":"2.0","info":{"version":"v1","title":"webpac","description":"a api to access plc data in a structured way"},"basePath":"/","paths":{"/api/Absolutes":{"get":{"tags":["Absolutes"],"operationId":"ApiAbsolutesGet","produces":["text/plain","application/json","text/json"],"parameters":[{"name":"Authorization","in":"header","description":"access token","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"type":"string"}}}},"deprecated":false}},"/api/Absolutes/{area}/{address}":{"get":{"tags":["Absolutes"],"operationId":"ApiAbsolutesByAreaByAddressGet","produces":["text/plain","application/json","text/json"],"parameters":[{"name":"area","in":"path","required":true,"type":"string"},{"name":"address","in":"path","required":true,"type":"string"},{"name":"Authorization","in":"header","description":"access token","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}}},"deprecated":false},"patch":{"tags":["Absolutes"],"operationId":"ApiAbsolutesByAreaByAddressPatch","produces":[],"parameters":[{"name":"area","in":"path","required":true,"type":"string"},{"name":"address","in":"path","required":true,"type":"string"},{"name":"value","in":"body","required":false,"schema":{"type":"object"}},{"name":"Authorization","in":"header","description":"access token","required":true,"type":"string"}],"responses":{"204":{"description":"No Content"}},"deprecated":false}},"/api/Absolutes/{area}":{"get":{"tags":["Absolutes"],"operationId":"ApiAbsolutesByAreaGet","produces":["text/plain","application/json","text/json"],"parameters":[{"name":"area","in":"path","required":true,"type":"string"},{"name":"addresses","in":"body","required":false,"schema":{"type":"array","items":{"type":"string"}}},{"name":"Authorization","in":"header","description":"access token","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}}},"deprecated":false},"patch":{"tags":["Absolutes"],"operationId":"ApiAbsolutesByAreaPatch","produces":[],"parameters":[{"name":"area","in":"path","required":true,"type":"string"},{"name":"value","in":"body","required":false,"schema":{"type":"object","additionalProperties":{"type":"object"}}},{"name":"Authorization","in":"header","description":"access token","required":true,"type":"string"}],"responses":{"204":{"description":"No Content"}},"deprecated":false}},"/api/Symbolic":{"get":{"tags":["Symbolic"],"operationId":"ApiSymbolicGet","produces":["text/plain","application/json","text/json"],"parameters":[{"name":"Authorization","in":"header","description":"access token","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"type":"string"}}}},"deprecated":false}},"/api/Symbolic/{mapping}/{variable}":{"get":{"tags":["Symbolic"],"operationId":"ApiSymbolicByMappingByVariableGet","produces":["text/plain","application/json","text/json"],"parameters":[{"name":"mapping","in":"path","required":true,"type":"string"},{"name":"variable","in":"path","required":true,"type":"string"},{"name":"Authorization","in":"header","description":"access token","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}}},"deprecated":false},"patch":{"tags":["Symbolic"],"operationId":"ApiSymbolicByMappingByVariablePatch","produces":[],"parameters":[{"name":"mapping","in":"path","required":true,"type":"string"},{"name":"variable","in":"path","required":true,"type":"string"},{"name":"value","in":"body","required":false,"schema":{"type":"object"}},{"name":"Authorization","in":"header","description":"access token","required":true,"type":"string"}],"responses":{"204":{"description":"No Content"}},"deprecated":false}},"/api/Symbolic/{mapping}":{"get":{"tags":["Symbolic"],"operationId":"ApiSymbolicByMappingGet","produces":["text/plain","application/json","text/json"],"parameters":[{"name":"mapping","in":"path","required":true,"type":"string"},{"name":"variables","in":"body","required":false,"schema":{"type":"array","items":{"type":"string"}}},{"name":"Authorization","in":"header","description":"access token","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}}},"deprecated":false},"put":{"tags":["Symbolic"],"operationId":"ApiSymbolicByMappingPut","produces":[],"parameters":[{"name":"mapping","in":"path","required":true,"type":"string"},{"name":"value","in":"body","required":false,"schema":{"type":"object"}},{"name":"Authorization","in":"header","description":"access token","required":true,"type":"string"}],"responses":{"204":{"description":"No Content"}},"deprecated":false},"patch":{"tags":["Symbolic"],"operationId":"ApiSymbolicByMappingPatch","produces":[],"parameters":[{"name":"mapping","in":"path","required":true,"type":"string"},{"name":"value","in":"body","required":false,"schema":{"type":"object","additionalProperties":{"type":"object"}}},{"name":"Authorization","in":"header","description":"access token","required":true,"type":"string"}],"responses":{"204":{"description":"No Content"}},"deprecated":false}},"/api/Token":{"post":{"tags":["Token"],"operationId":"ApiTokenPost","produces":["text/plain","application/json","text/json"],"parameters":[{"name":"req","in":"body","required":false,"schema":{"$ref":"#/definitions/AuthRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Authentication"}}},"deprecated":false}}},"definitions":{"AuthRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"Authentication":{"type":"object","properties":{"Authenticated":{"type":"boolean"},"User":{"type":"string"},"Role":{"type":"string"},"Token":{"type":"string"},"TokenExpires":{"format":"date-time","type":"string"}}}},"securityDefinitions":{}}